requirements.txt
openai
requests
python-dotenv
slack-sdk>=3,<4
slackeventsapi>=3,<4
lru-dict